Fix scanner incremental scan

+++ src/core-impl/collections/db/ScanResultProcessor.cpp
@@ -74,7 +74,8 @@ ScanResultProcessor::commit()
 
     // -- now check if some of the tracks are not longer used and also not moved to another directory
     foreach( const CollectionScanner::Directory* dir, m_directories )
-        deleteDeletedTracks( getDirectory( dir->path(), dir->mtime() ) );
+        if( !dir->isSkipped() )
+            deleteDeletedTracks( getDirectory( dir->path(), dir->mtime() ) );
 
     // -- delete all not-found directories
     if( m_type != PartialUpdateScan )
